DevelopCommand: try to make it work on iOS
